src/clients/desktop_app/ami_daemon/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
httpx>=0.25.0
pyyaml>=6.0.0
websockets>=12.0
python-multipart>=0.0.6
psutil>=5.9.0
aiosqlite>=0.19.0
anthropic>=0.25.0
openai>=1.0.0
browser-use>=0.1.0
playwright>=1.40.0
pathspec>=0.11.0
claude-agent-sdk>=0.1.0
pyinstaller>=6.0.0
